Change usage of `.len()` to `.chars().count()`
This fixes some display problems with teamnames that contain unicode
symbols beyond ascii.
I wrapped the calls to `strip_ansi_escapes` in their own function now,
so that it can return a proper string rather than a vector of bytes, and
we can actually call `.chars()` on it.

Had to change some `+ 1` to a `- 1` to make the alignment work again,
I'm not quite sure why, but I assume it might be because the unicode box
symbols now also only count as a single character, rather than the
number of bytes they require in utf8.

This of course assumes things are utf8 encoded, but do we really care
about anything else :)